requirements.txt
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.3.1
environs==9.5.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
marshmallow==3.20.1
packaging==23.2
python-dotenv==1.0.0
python-telegram-bot==20.6
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.7
